Move back to using type assertions
Using type assertions seems to be better go than
using a tagged struct, so deleted the Data struct and renmaed
Item -> Data.

Also added a method to the Data instance to ensure that only
our new types are usable as Data, so we get better compile-time
verification that we're using the correct types was Data. Currently
that method is just String(), and will aim to keep that as simple
as possible.